+++ packages/modeling/src/geometries/path2/concat.js
@@ -1,8 +1,11 @@
 const fromPoints = require('./fromPoints')
 const toPoints = require('./toPoints')
+
 const { equals } = require('../../maths/vec2')
+
 /**
  * Concatenate the given paths.
+ *
  * If both contain the same point at the junction, merge it into one.
  * A concatenation of zero paths is an empty, open path.
  * A concatenation of one closed path to a series of open paths produces a closed path.
@@ -17,16 +20,14 @@ const { equals } = require('../../maths/vec2')
 const concat = (...paths) => {
   // Only the last path can be closed, producing a closed path.
   let isClosed = false
-  for (const path of paths) {
-    if (isClosed) {
-      throw new Error('Cannot concatenate to a closed path')
+  let newpoints = []
+  paths.forEach((path, i) => {
+    const tmp = toPoints(path).slice()
+    if (newpoints.length > 0 && tmp.length > 0 && equals(tmp[0], newpoints[newpoints.length - 1])) tmp.shift()
+    if (tmp.length > 0 && isClosed) {
+      throw new Error(`Cannot concatenate to a closed path; check the ${i}th path`)
     }
     isClosed = path.isClosed
-  }
-  let newpoints = []
-  paths.forEach((path) => {
-    const tmp = toPoints(path)
-    if (newpoints.length > 0 && tmp.length > 0 && equals(tmp[0], newpoints[newpoints.length - 1])) tmp.shift()
     newpoints = newpoints.concat(tmp)
   })
   return fromPoints({ closed: isClosed }, newpoints)