+++ packages/modeling/src/maths/utils/trigonometry.js
@@ -6,28 +6,28 @@ const { NEPS } = require('../constants')
 const rezero = (n) => Math.abs(n) < NEPS ? 0 : n
 
 /**
- * Return Math.sin but accurate for 90 degree rotations.
+ * Return Math.sin but accurate for TAU / 4 rotations.
  * Fixes rounding errors when sin should be 0.
  *
  * @param {Number} radians - angle in radians
  * @returns {Number} sine of the given angle
  * @alias module:modeling/utils.sin
  * @example
- * sin(Math.PI) == 0
- * sin(2 * Math.PI) == 0
+ * sin(TAU / 2) == 0
+ * sin(TAU) == 0
  */
 const sin = (radians) => rezero(Math.sin(radians))
 
 /**
- * Return Math.cos but accurate for 90 degree rotations.
+ * Return Math.cos but accurate for TAU / 4 rotations.
  * Fixes rounding errors when cos should be 0.
  *
  * @param {Number} radians - angle in radians
  * @returns {Number} cosine of the given angle
  * @alias module:modeling/utils.cos
  * @example
- * cos(0.5 * Math.PI) == 0
- * cos(1.5 * Math.PI) == 0
+ * cos(TAU * 0.25) == 0
+ * cos(TAU * 0.75) == 0
  */
 const cos = (radians) => rezero(Math.cos(radians))