+++ packages/modeling/src/utils/radToDeg.test.js
@@ -1,19 +1,19 @@
 const test = require('ava')
 
+const { TAU } = require('../maths/constants')
+
 const { radToDeg } = require('./index')
 
 test('utils: radToDeg() should return correct values', (t) => {
-  const radians = Math.PI / 180
-
-  const obs1 = radToDeg(radians * 0)
+  const obs1 = radToDeg(0)
   t.true(obs1 === 0)
 
-  const obs2 = radToDeg(radians * 90)
+  const obs2 = radToDeg(TAU / 4)
   t.true(obs2 === 90)
 
-  const obs3 = radToDeg(radians * 180)
+  const obs3 = radToDeg(TAU / 2)
   t.true(obs3 === 180)
 
-  const obs4 = radToDeg(radians * 270)
+  const obs4 = radToDeg(TAU * 0.75)
   t.true(obs4 === 270)
 })