+++ packages/io/x3d-serializer/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.4.2](https://github.com/jscad/OpenJSCAD.org/compare/@jscad/x3d-serializer@2.4.1...@jscad/x3d-serializer@2.4.2) (2022-08-21)
+
+**Note:** Version bump only for package @jscad/x3d-serializer
+
+
+
+
+
+## [2.4.1](https://github.com/jscad/OpenJSCAD.org/compare/@jscad/x3d-serializer@2.4.0...@jscad/x3d-serializer@2.4.1) (2022-07-17)
+
+**Note:** Version bump only for package @jscad/x3d-serializer
+
+
+
+
+
 # [2.4.0](https://github.com/jscad/OpenJSCAD.org/compare/@jscad/x3d-serializer@2.3.9...@jscad/x3d-serializer@2.4.0) (2022-06-12)